Reworking & reorganizing code
