+++ app/passport_score/tests/test_loading_stamps.py
@@ -4,8 +4,8 @@ from django.utils import timezone
 import pytest
 from dashboard.models import Passport, PassportStamp
 from passport_score.gr15_providers import providers
-from passport_score.management.commands.test_compute_gr15_trust_bonus import load_passport_stamps
 from passport_score.models import GR15TrustScore
+from passport_score.utils import load_passport_stamps
 from test_plus.test import TestCase
 
 NUM_USERS = 10
