+++ app/passport_score/utils.py
@@ -14,9 +14,29 @@ MAX_TRUST_BONUS = 1.5
 MIN_TRUST_BONUS = 0.5
 
 
+def get_apu_median():
+    """Calculate the median APU from the records in GR15TrustScore"""
+    df_gr15_calculations = pd.DataFrame(GR15TrustScore.objects.values("last_apu_score"))
+    return df_gr15_calculations.last_apu_score.median()
+
+
+def get_apu_min():
+    """Calculate the min APU from the records in GR15TrustScore"""
+    df_gr15_calculations = pd.DataFrame(GR15TrustScore.objects.values("last_apu_score"))
+    return df_gr15_calculations.last_apu_score.min()
+
+
 def load_passport_stamps():
     """
     Load stamps from the grants database & prepares the data for input into the GR15 scoring algorithm
+    Will return a dataframe like:
+        - 1 row for each user
+        - user_id is the index
+        - 1 row for each configured provider with 1 if the user has that stamp
+        - column `is_stamp_preserved`
+            - 1 - if all the stamps already in GR15TrustScore have been kept by the user
+            - 0 - if stamps from GR15TrustScore have been deleted by the user
+
     """
 
     # Load stamps and store them into a dataframe
@@ -121,9 +141,7 @@ def load_passport_stamps():
     # By applying min, we identify thoses users (as they end up having 0)
     aggregation_rule["is_stamp_preserved"] = min
     prepared_df = df_stamp_overview.groupby(["user_id"]).agg(aggregation_rule)
-    print("=" * 40)
     print("prepared_df\n", prepared_df)
-    print("=" * 40)
     prepared_df.reset_index(inplace=True)
 
     # Reset the index, we want to have the user_id as separate column
@@ -149,16 +167,20 @@ def calculate_trust_bonus(df_existing_gr15_scores, df_apu_scores):
     # new records from records to be updated)
     df_existing_gr15_scores["from_db"] = True
 
-    apu_median = df_apu_scores["Score"].median()
-    apu_min = df_apu_scores["Score"].min()
-    print("APU median: %s" % apu_median)
-    print("APU min: %s" % apu_min)
-
     df_gr15_scores = df_existing_gr15_scores.join(df_apu_scores, how="outer")
 
     # Ensure valid values in Score. NaN might occur for users that have removed all thei stamps
     df_gr15_scores.Score.fillna(0.0, inplace=True)
-    df_gr15_scores.has_removed_stamps.fillna(False, inplace=True)
+    print("Score:\n", df_gr15_scores.Score)
+
+    apu_median = df_gr15_scores.Score.median()
+    apu_min = df_gr15_scores.Score.min()
+    print("APU median: %s" % apu_median)
+    print("APU min: %s" % apu_min)
+
+    # NA in has_removed_stamps means all stamps have been reduced, otherwise this value would
+    # a valid True / False value ...
+    df_gr15_scores.has_removed_stamps.fillna(True, inplace=True)
     df_gr15_scores.current_stamps.fillna("", inplace=True)
 
     print("\nInitial scores summary (including db records): \n%s" % df_gr15_scores)
@@ -167,9 +189,6 @@ def calculate_trust_bonus(df_existing_gr15_scores, df_apu_scores):
     now = timezone.now()
     df_gr15_scores.last_apu_score = df_gr15_scores.Score
     df_gr15_scores.last_apu_calculation_time = now
-
-    print("last_apu_score:", df_gr15_scores.last_apu_score)
-    print("max_apu_score:", df_gr15_scores.max_apu_score)
 
     max_apus_to_update = (pd.isnull(df_gr15_scores.max_apu_score)) | (
         df_gr15_scores.last_apu_score > df_gr15_scores.max_apu_score
@@ -199,15 +218,9 @@ def calculate_trust_bonus(df_existing_gr15_scores, df_apu_scores):
     new_trust_bonus.clip(upper=1.5, inplace=True)
 
     # Ensure valid values (eliminate NaNs)
-    new_trust_bonus.fillna(0.0, inplace=True)
+    new_trust_bonus.fillna(MAX_TRUST_BONUS, inplace=True)
     new_trust_bonus.replace([np.inf, -np.inf], 0, inplace=True)
     df_gr15_scores.trust_bonus.fillna(0.0, inplace=True)
-
-    print("-----------------")
-    print("new_trust_bonus", new_trust_bonus)
-    print("-----------------")
-    print("df_gr15_scores.trust_bonus", df_gr15_scores.trust_bonus)
-    print("-----------------")
 
     update_trust_bonus_condition = (df_gr15_scores.has_removed_stamps == True) | (
         new_trust_bonus > df_gr15_scores.trust_bonus
@@ -229,172 +242,169 @@ def calculate_trust_bonus(df_existing_gr15_scores, df_apu_scores):
     return df_gr15_scores
 
 
-class Command(BaseCommand):
-    help = "Calculates the APU score for GR15 based on the currently submitted passports & stamps"
+def get_new_trust_bonus():
 
-    def handle(self, *args, **options):
-        # delay import as this is only available in newer envs ...
-        from passport_score.gr15_scorer import compute_apu_scores
+    # delay import as this is only available in newer envs ...
+    from passport_score.gr15_scorer import compute_apu_scores
 
-        try:
-            ###################################################################################
-            # Load the data and compute the APU score
-            ###################################################################################
+    try:
+        ###################################################################################
+        # Load the data and compute the APU score
+        ###################################################################################
 
-            self.stdout.write("Loading passport stamps")
+        print("Loading passport stamps")
 
-            (
-                prepared_df,
-                stamp_fields,
-                grouping_fields,
-                grouping_fields_1,
-            ) = load_passport_stamps()
+        (
+            prepared_df,
+            stamp_fields,
+            grouping_fields,
+            grouping_fields_1,
+        ) = load_passport_stamps()
 
-            self.stdout.write("Running scorer")
-            df_apu_scores = compute_apu_scores(
-                prepared_df, stamp_fields, grouping_fields, grouping_fields_1
-            )
+        print("Running scorer")
+        df_apu_scores = compute_apu_scores(
+            prepared_df, stamp_fields, grouping_fields, grouping_fields_1
+        )
 
-            df_apu_scores.set_index("user_id", inplace=True)
-            prepared_df.set_index("user_id", inplace=True)
-            # This will initialize all NaNs with []  (see https://stackoverflow.com/a/64207857)
-            print("===>df_apu_scores\n", df_apu_scores)
-            print("===>prepared_df\n", prepared_df)
+        df_apu_scores.set_index("user_id", inplace=True)
+        prepared_df.set_index("user_id", inplace=True)
+        # This will initialize all NaNs with []  (see https://stackoverflow.com/a/64207857)
+        print("===>df_apu_scores\n", df_apu_scores)
+        print("===>prepared_df\n", prepared_df)
 
-            df_t = pd.DataFrame(
+        df_t = pd.DataFrame(
+            {
+                "current_stamps": prepared_df.stamp_provider,
+                "has_removed_stamps": (1 - prepared_df.is_stamp_preserved).astype(
+                    "bool"
+                ),
+            },
+            index=prepared_df.index,
+        )
+        print("===>df_t\n", df_t)
+
+        df_apu_scores = df_apu_scores.merge(
+            df_t, how="outer", left_index=True, right_index=True
+        )
+
+        # df_apu_scores["current_stamps"] = prepared_df.stamp_provider
+        # df_apu_scores["current_stamps"].fillna("HELLO", inplace=True)
+        # df_apu_scores["has_removed_stamps"] = (
+        #     1 - prepared_df.is_stamp_preserved
+        # )  # Just invert the logic here
+        print("\nScores: \n%s" % df_apu_scores)
+
+        ###################################################################################
+        # Evaluate the APU score, apply the trust bonus calculation
+        ###################################################################################
+        # Load existing scores from the DB
+        existing_gr15_scores = GR15TrustScore.objects.values(
+            "id",
+            "user_id",
+            "last_apu_score",
+            "max_apu_score",
+            "trust_bonus",
+            "last_apu_calculation_time",
+            "max_apu_calculation_time",
+            "trust_bonus_calculation_time",
+            "stamps",
+        )
+
+        # Create a DataFrame for the data, handle the case when the dataset loaded from DB is empty
+        if existing_gr15_scores:
+            # Create dataframe from DB results
+            df_existing_gr15_scores = pd.DataFrame(existing_gr15_scores)
+        else:
+            # Create an empty dataframe, if no results where loaded from DB
+            df_existing_gr15_scores = pd.DataFrame.from_dict(
                 {
-                    "current_stamps": prepared_df.stamp_provider,
-                    "has_removed_stamps": (1 - prepared_df.is_stamp_preserved).astype(
-                        "bool"
-                    ),
-                },
-                index=prepared_df.index,
-            )
-            print("===>df_t\n", df_t)
-
-            df_apu_scores = df_apu_scores.merge(
-                df_t, how="outer", left_index=True, right_index=True
-            )
-
-            # df_apu_scores["current_stamps"] = prepared_df.stamp_provider
-            # df_apu_scores["current_stamps"].fillna("HELLO", inplace=True)
-            # df_apu_scores["has_removed_stamps"] = (
-            #     1 - prepared_df.is_stamp_preserved
-            # )  # Just invert the logic here
-            self.stdout.write("\nScores: \n%s" % df_apu_scores)
-
-            ###################################################################################
-            # Evaluate the APU score, apply the trust bonus calculation
-            ###################################################################################
-            # Load existing scores from the DB
-            existing_gr15_scores = GR15TrustScore.objects.values(
-                "id",
-                "user_id",
-                "last_apu_score",
-                "max_apu_score",
-                "trust_bonus",
-                "last_apu_calculation_time",
-                "max_apu_calculation_time",
-                "trust_bonus_calculation_time",
-                "stamps",
+                    "id": [],
+                    "user_id": [],
+                    "last_apu_score": [],
+                    "max_apu_score": [],
+                    "trust_bonus": [],
+                    "last_apu_calculation_time": [],
+                    "max_apu_calculation_time": [],
+                    "trust_bonus_calculation_time": [],
+                    "stamps": [],
+                }
             )
 
-            # Create a DataFrame for the data, handle the case when the dataset loaded from DB is empty
-            if existing_gr15_scores:
-                # Create dataframe from DB results
-                df_existing_gr15_scores = pd.DataFrame(existing_gr15_scores)
-            else:
-                # Create an empty dataframe, if no results where loaded from DB
-                df_existing_gr15_scores = pd.DataFrame.from_dict(
-                    {
-                        "id": [],
-                        "user_id": [],
-                        "last_apu_score": [],
-                        "max_apu_score": [],
-                        "trust_bonus": [],
-                        "last_apu_calculation_time": [],
-                        "max_apu_calculation_time": [],
-                        "trust_bonus_calculation_time": [],
-                        "stamps": [[]],
-                    }
-                )
+        df_existing_gr15_scores.set_index("user_id", inplace=True)
+        df_existing_gr15_scores[
+            "original_last_apu_score"
+        ] = df_existing_gr15_scores.last_apu_score
+        df_existing_gr15_scores[
+            "original_trust_bonus"
+        ] = df_existing_gr15_scores.trust_bonus
+        df_gr15_scores = calculate_trust_bonus(df_existing_gr15_scores, df_apu_scores)
 
-            df_existing_gr15_scores.set_index("user_id", inplace=True)
-            df_existing_gr15_scores["original_last_apu_score"] = df_existing_gr15_scores.last_apu_score
-            df_existing_gr15_scores["original_trust_bonus"] = df_existing_gr15_scores.trust_bonus
-            df_gr15_scores = calculate_trust_bonus(
-                df_existing_gr15_scores, df_apu_scores
-            )
+        return df_gr15_scores
+    except Exception as exc:
+        print('ERROR: "%s"' % exc)
+        # printing stack trace
+        traceback.print_exc()
+        raise CommandError("An unexpected error occured")
 
-            # Drop any records with no current_stamps
-            # df_gr15_scores.dropna(subset=["current_stamps"], inplace=True)
-            self.stdout.write("\nNew scores & trust bonus: \n%s" % df_gr15_scores)
-            ###################################################################################
-            # Write the new scores back to the DB
-            ###################################################################################
-            # Determine the records to be created in the DB (for new users that have submitted their passport) and the ones to be created
-            new_records = [
-                GR15TrustScore(
-                    user_id=user_id,
-                    last_apu_score=data.last_apu_score,
-                    max_apu_score=data.max_apu_score,
-                    trust_bonus=data.trust_bonus,
-                    last_apu_calculation_time=data.last_apu_calculation_time,
-                    max_apu_calculation_time=data.max_apu_calculation_time,
-                    trust_bonus_calculation_time=data.trust_bonus_calculation_time,
-                    stamps=data.current_stamps,
-                )
-                for user_id, data in df_gr15_scores.loc[
-                    df_gr15_scores.from_db != True
-                ].iterrows()
-            ]
 
-            records_from_db = [
-                GR15TrustScore(
-                    id=data.id,
-                    user_id=user_id,
-                    last_apu_score=data.last_apu_score,
-                    max_apu_score=data.max_apu_score,
-                    trust_bonus=data.trust_bonus,
-                    last_apu_calculation_time=data.last_apu_calculation_time,
-                    max_apu_calculation_time=data.max_apu_calculation_time,
-                    trust_bonus_calculation_time=data.trust_bonus_calculation_time,
-                    stamps=data.current_stamps,
-                )
-                for user_id, data in df_gr15_scores.loc[
-                    df_gr15_scores.from_db == True
-                ].iterrows()
-            ]
+def save_gr15_trustbonus_records(df_gr15_scores):
+    ###################################################################################
+    # Write the new scores back to the DB
+    ###################################################################################
+    # Determine the records to be created in the DB (for new users that have submitted their passport) and the ones to be created
+    new_records = [
+        GR15TrustScore(
+            user_id=user_id,
+            last_apu_score=data.last_apu_score,
+            max_apu_score=data.max_apu_score,
+            trust_bonus=data.trust_bonus,
+            last_apu_calculation_time=data.last_apu_calculation_time,
+            max_apu_calculation_time=data.max_apu_calculation_time,
+            trust_bonus_calculation_time=data.trust_bonus_calculation_time,
+            stamps=data.current_stamps,
+        )
+        for user_id, data in df_gr15_scores.loc[
+            df_gr15_scores.from_db != True
+        ].iterrows()
+    ]
 
-            self.stdout.write("\nRecords to be updated: \n%s" % records_from_db)
-            self.stdout.write("\nNew records to be created: \n%s" % new_records)
+    records_from_db = [
+        GR15TrustScore(
+            id=data.id,
+            user_id=user_id,
+            last_apu_score=data.last_apu_score,
+            max_apu_score=data.max_apu_score,
+            trust_bonus=data.trust_bonus,
+            last_apu_calculation_time=data.last_apu_calculation_time,
+            max_apu_calculation_time=data.max_apu_calculation_time,
+            trust_bonus_calculation_time=data.trust_bonus_calculation_time,
+            stamps=data.current_stamps,
+        )
+        for user_id, data in df_gr15_scores.loc[
+            df_gr15_scores.from_db == True
+        ].iterrows()
+    ]
 
-            # Bulk creating new records
-            self.stdout.write(
-                "\nBulk creating new records (count=%s)\n" % len(new_records)
-            )
-            GR15TrustScore.objects.bulk_create(new_records)
+    # print("\nRecords to be updated: \n%s" % records_from_db)
+    print("\nNew records to be created: \n%s" % new_records)
 
-            # Bulk updating existing records
-            self.stdout.write(
-                "\nBulk updating existing records (count=%s)\n" % len(records_from_db)
-            )
+    # Bulk creating new records
+    print("\nBulk creating new records (count=%s)\n" % len(new_records))
+    GR15TrustScore.objects.bulk_create(new_records)
 
-            GR15TrustScore.objects.bulk_update(
-                records_from_db,
-                [
-                    "user_id",
-                    "last_apu_score",
-                    "max_apu_score",
-                    "trust_bonus",
-                    "last_apu_calculation_time",
-                    "max_apu_calculation_time",
-                    "trust_bonus_calculation_time",
-                    "stamps",
-                ],
-            )
-        except Exception as exc:
-            self.stdout.write(self.style.ERROR('ERROR: "%s"' % exc))
-            # printing stack trace
-            traceback.print_exc()
-            raise CommandError("An unexpected error occured")
+    # Bulk updating existing records
+    print("\nBulk updating existing records (count=%s)\n" % len(records_from_db))
+
+    GR15TrustScore.objects.bulk_update(
+        records_from_db,
+        [
+            "user_id",
+            "last_apu_score",
+            "max_apu_score",
+            "trust_bonus",
+            "last_apu_calculation_time",
+            "max_apu_calculation_time",
+            "trust_bonus_calculation_time",
+            "stamps",
+        ],
+    )
